Use alternative dos if available

+++ floppy_gen.sh
@@ -1,5 +1,11 @@
 #!/bin/sh
 
-cp boot_floppy/FLOPPY.img tandy_game.img
+if test -f boot_floppy/Dos3.3.img
+then
+    cp boot_floppy/Dos3.3.img tandy_game.img
+else
+    cp boot_floppy/FLOPPY.img tandy_game.img
+fi
+
 mcopy -i tandy_game.img release ::game
 mcopy -o -i tandy_game.img boot_floppy/autoexec.bat ::AUTOEXEC.BAT
